+++ src/main/java/sender/util/JSONParser/AddressBookJsonParser.java
@@ -13,7 +13,7 @@ public class AddressBookJsonParser implements JsonParser<Addressbook> {
     public Addressbook parseOne(String json) {
 
         String[] arr = json.split(",");
-        int id = new Integer(arr[0].substring(27));
+        long id = new Long(arr[0].substring(27));
         String name = arr[1].substring(15, arr[1].length() - 1);
         int phones = new Integer(arr[2].substring(16));
         int exceptions = new Integer(arr[3].substring(20));
@@ -34,7 +34,7 @@ public class AddressBookJsonParser implements JsonParser<Addressbook> {
         for (int i = 2; i < objects.length; i++) {
 
             String[] fields = objects[i].split("\n");
-            int id = new Integer(fields[1].substring(13, fields[1].length() - 2));
+            long id = new Long(fields[1].substring(13, fields[1].length() - 2));
             String name = fields[2].substring(13, fields[2].length() - 2);
             int phone = new Integer(fields[3].substring(13, fields[3].length() - 2));
             String date = fields[4].substring(13, fields[4].length() - 2);
